+++ RELEASE_NOTES.rst
@@ -8,7 +8,7 @@ Release notes
   messages.
 * ``autohelp.py``: Improve sanitizer, better support for i18n in projects.
 * ``autohelp-wrapper``: Smarter handling of the manuals repo and environment
-  setup.
+  setup, add support for the ``create`` subcommand.
 * ``diff_branches.py``: Updated output format.
 
 0.18.1