+++ RELEASE_NOTES.rst
@@ -9,7 +9,7 @@ Release notes
 * ``autohelp.py``: Improve sanitizer, better support for i18n in
   projects, allow setting of title name for tables.
 * ``autohelp-wrapper``: Smarter handling of the manuals repo and environment
-  setup.
+  setup, add support for the ``create`` subcommand.
 * ``diff_branches.py``: Updated output format.
 
 0.18.1