+++ RELEASE_NOTES.rst
@@ -15,7 +15,8 @@ Release notes
 * ``diff_branches.py``: Updated output format.
 * Provide a ``hotref`` extension for sphinx, to automate the creation of
   references to the HOT resources documentation.
-* ``openstack-auto-commands``: Handle python-openstackclient.
+* ``openstack-auto-commands``: Handle python-openstackclient, handle
+  python-glanceclient v2 commands.
 
 0.18.1
 ------